requirements.txt
django~=4.2.3
gunicorn==21.2.0
django-braces~=1.15.0
djangorestframework~=3.14.0
drf-spectacular==0.26.4
drf-yasg~=1.21.7